-some minor changes to Namelist nh_pzlev - bug fixes for zlevels and plevels: INTEGER -> REAL

+++ src/configure_model/mo_nh_pzlev_config.f90
@@ -67,9 +67,9 @@ MODULE mo_nh_pzlev_config
 
     INTEGER :: nplev                 !< number of p-levels
 
-    INTEGER :: zlevels(100)          !< zlevel heights [m] 
+    REAL(wp):: zlevels(100)          !< zlevel heights [m] 
 
-    INTEGER :: plevels(100)          !< plevel heights [m] 
+    REAL(wp):: plevels(100)          !< plevel heights [m] 
 
     ! derived variables
     !
